requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pytrends==4.9.2
wikipedia==1.4.0
Werkzeug==3.0.1
pandas==2.1.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
langcodes==3.5.0
openai==1.43.0
markdown==3.6
google-api-python-client==2.129.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
stripe==10.11.0
